% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_search.R
\name{cluster_search}
\alias{cluster_search}
\title{Search of heirachical clustering}
\usage{
cluster_search(hClust, varMat, sizeMin)
}
\arguments{
\item{hClust}{Heirachical clustering.}

\item{varMat}{Variance matrix}

\item{sizeMin}{Minimum size of cluster}
}
\value{
A list of results:
\itemize{
\item{features}: A boolean matrix of features with each row corresponding to each tree cut.
\item{value}: The average value of the corresponding set of features.
\item{size}: The number of features selected in each cut.
}
}
\description{
Search of heirachical clustering
}
