% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inf_criterion.R
\name{inf_criterion}
\alias{inf_criterion}
\title{Information Criterion}
\usage{
inf_criterion(
  model,
  x,
  y,
  family = "binomial",
  lambda = "lambda.1se",
  metric = "BIC",
  gamma = gamma
)
}
\arguments{
\item{model}{Regression model.}

\item{x}{input matrix, rows represent observations and columns represent
features. column names and row names are recommended.}

\item{y}{response variable, restricted to two classes.}

\item{family}{GLM family.}

\item{lambda}{Used in cv.glmnet.}

\item{metric}{AIC,  AICc, BIC or EBIC.}

\item{gamma}{value for EBIC.}
}
\value{
The value of the metric when calculated from the model.
}
\description{
Information Criterion
}
\examples{

library('ropls')
data("sacurine") #Load sacurine dataset from the 'ropls' package

dat = sacurine$dataMatrix
outcomes = sacurine$sampleMetadata$gender

inf_criterion(model = rep(TRUE, NCOL(dat)), x = dat, y = outcomes)
}
