% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vivid.R
\name{vivid}
\alias{vivid}
\title{VIVID}
\usage{
vivid(
  x,
  y,
  bootstraps = 100,
  cores = parallel::detectCores() - 1,
  crossfold = 10,
  seed = 1234567,
  minSize = 2,
  lambda = "lambda.1se",
  compareMethod = "BIC",
  gamma = 1
)
}
\arguments{
\item{x}{input matrix, rows represent observations and columns represent
features. column names and row names are recommended.}

\item{y}{response variable, restricted to two classes.}

\item{bootstraps}{integer, the number of bootstraps to run.}

\item{cores}{integer, number of cores to use for parallel processing. VIVID
makes uses of the "future" and "furrr" packages and will set the plan to
"multiprocess". The cores parameter is directly fed to workers arguement of
future::plan.}

\item{crossfold}{Integer number of cross folds used.}

\item{seed}{integer, passed to set.seed function, a full L'Ecuyer-CMRG RNG seed (vector of 7 integers).}

\item{minSize}{minimum cluster size to consider when clustering.}

\item{lambda}{Used in cv.glmnet.}

\item{compareMethod}{AIC,  AICc, BIC or EBIC.}

\item{gamma}{gamma value for EBIC.}
}
\value{
A list of results:
\itemize{
\item{coefficients}: A numeric matrix of estimated coefficient values over re-samples.
\item{varMat}: A numeric matrix of the estimated variance matrix explain in the VIVID method.
\item{varClust}: The hclust() output of the clustering performed in the VIVID method.
\item{selection}: A boolean matrix where each row indicates a model identified as a potential final model.
\item{sizes}: A integer vector of model sizes in the seleciton path.
\item{compareMethod}: A character value indicating the Information Criterion used for the results.
\item{compareValues}: A numeric vector of Infomation Criterion values for each model in the selection path.
\item{optModel}: A boolean vector indicating whether or not each feature is include in the final model.
\item{optFeatures}: A character vector containing all features in the final model.
\item{vividSplit}: A boolean value indicating whether or not this output is from the vivid_split() funciton.
}
}
\description{
VIVID
}
