% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vivid_crit.R
\name{vivid_crit}
\alias{vivid_crit}
\title{Change the information criteria used in VIVID}
\usage{
vivid_crit(vividObj, x, y, metric)
}
\arguments{
\item{vividObj}{An object passed from the vivid() funciton.}

\item{x}{input matrix, rows represent observations and columns represent
features. column names and row names are recommended.}

\item{y}{response variable, restricted to two classes.}

\item{metric}{AIC,  AICc, BIC or EBIC.}
}
\value{
A list of results:
\itemize{
\item{coefficients}: A numeric matrix of estimated coefficient values over re-samples.
\item{varMat}: A numeric matrix of the estimated variance matrix explain in the VIVID method.
\item{varClust}: The hclust() output of the clustering performed in the VIVID method.
\item{selection}: A boolean matrix where each row indicates a model identified as a potential final model.
\item{sizes}: A integer vector of model sizes in the seleciton path.
\item{compareMethod}: A character value indicating the Information Criterion used for the results.
\item{compareValues}: A numeric vector of Infomation Criterion values for each model in the selection path.
\item{optModel}: A boolean vector indicating whether or not each feature is include in the final model.
\item{optFeatures}: A character vector containing all features in the final model.
\item{vividSplit}: A boolean value indicating whether or not this output is from the vivid_split() funciton.
}
}
\description{
Change the information criteria used in VIVID
}
\examples{
library('ropls')
data("sacurine") #Load sacurine dataset from the 'ropls' package

dat = sacurine$dataMatrix
outcomes = sacurine$sampleMetadata$gender

vividResults = vivid(x = dat,
      y = outcomes)
      
newVivid = vivid_crit(vividObj = vividResults, x = dat, y = outcomes, metric = 'AIC')

newVivid$optFeatures
}
