% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vivid_reg.R
\name{vivid_reg}
\alias{vivid_reg}
\title{Regression method used in the VIVID method}
\usage{
vivid_reg(weight, x, y, crossfold = 10, lambda = "lambda.1se")
}
\arguments{
\item{weight}{A p length vector}

\item{x}{input matrix, rows represent observations and columns represent
features. column names and row names are recommended.}

\item{y}{response variable, restricted to two classes.}

\item{crossfold}{Number of crossfold samples.}

\item{lambda}{Used in cv.glmnet.}
}
\value{
A vector of estimated coefficient values.
}
\description{
Regression method used in the VIVID method
}
\examples{
library('ropls')
data("sacurine") #Load sacurine dataset from the 'ropls' package

dat = sacurine$dataMatrix
outcomes = sacurine$sampleMetadata$gender

vivid_reg(weight = rep(1, NROW(dat)), x = dat, y = outcomes)
}
