\name{VIVID-package}
\alias{VIVID-package}
\alias{VIVID}
\docType{package}
\title{
\packageTitle{VIVID}
}
\description{
\packageDescription{VIVID}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{VIVID}
\packageIndices{VIVID}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{VIVID}

Maintainer: \packageMaintainer{VIVID}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
